Fixes #25264 - Remove foreman_ansible_core dependency
This takes all of the foreman_ansible_core features and removes them.
These should only be used by smart_proxy_dynflow_core, so a new release
should be made where installing foreman_ansible_core only installs it
for smart_proxy_dynflow_core and not for the regular proxy.

I've also moved the tests for RolesReader from foreman_ansible to this
repository.
