+++ .travis.yml
@@ -1,0 +1,5 @@
+language: ruby
+rvm:
+  - 2.3
+  - 2.4
+  - 2.5
