+++ Gemfile
@@ -3,7 +3,13 @@ source 'https://rubygems.org'
 gemspec
 
 group :development do
-  #gem 'smart_proxy', :git => "https://github.com/theforeman/smart-proxy", :branch => "develop"
-  gem 'smart_proxy', :path => '../smart-proxy'
+  gem 'smart_proxy', git: 'https://github.com/theforeman/smart-proxy',
+                     branch: 'develop'
+  #gem 'smart_proxy', path: '../smart-proxy'
   gem 'pry'
+end
+
+group :test do
+  gem 'minitest'
+  gem 'mocha'
 end
