+++ Rakefile
@@ -1,11 +1,11 @@
 require 'rake'
 require 'rake/testtask'
-require "bundler/gem_tasks"
+require 'bundler/gem_tasks'
 
 desc 'Default: run unit tests.'
-task :default => :test
+task default: :test
 
-desc 'Test Ssh plugin.'
+desc 'Test Ansible plugin'
 Rake::TestTask.new(:test) do |t|
   t.libs << '.'
   t.libs << 'lib'
