--- lib/foreman_ansible_core/version.rb
@@ -1,3 +1,0 @@
-module ForemanAnsibleCore
-  VERSION = '2.0.2'.freeze
-end
