+++ lib/smart_proxy_ansible.rb
@@ -4,5 +4,6 @@ module Proxy
   module Ansible
     require 'smart_proxy_ansible/version'
     require 'smart_proxy_ansible/plugin'
+    require 'smart_proxy_ansible/roles_reader'
   end
 end
